+++ fabric/src/main/java/de/crafty/eiv/fabric/integration/ModMenuIntegration.java
@@ -1,0 +1,14 @@
+package de.crafty.eiv.fabric.integration;
+
+import com.terraformersmc.modmenu.ModMenu;
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+import de.crafty.eiv.common.gui.EivClientSettingsScreen;
+import net.fabricmc.loader.api.FabricLoader;
+
+public class ModMenuIntegration implements ModMenuApi {
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        return (EivClientSettingsScreen::new);
+    }
+}