+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Added buttons to change the current page.
 - Added a "Search ItemView" hint to the search bar.
 - Item View position can now be swapped between the left and right panels.
+- EIV's settings can now be modified in Mod Menu.